extract workspace + redirect after default course creation
